requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiohttp
pydantic
python-dotenv
requests